e2e/requirements.txt
requests
requests-toolbelt
psycopg2-binary